5.0.1rel fix codecheck 20240925 update001 Signed-off-by heguangyao <heguangyao1@h-partners.com>

+++ frameworks/innerkitsimpl/accessor/src/abstract_exif_metadata_accessor.cpp
@@ -75,7 +75,7 @@ uint32_t AbstractExifMetadataAccessor::GetFilterArea(const std::vector<std::stri
         return E_NO_EXIF_TAG;
     }
     for (auto& range : ranges) {
-        range.first += GetTiffOffset();
+        range.first += static_cast<uint32_t>(GetTiffOffset());
     }
     return SUCCESS;
 }
